Improve handshake validation

+++ common/packet.go
@@ -60,11 +60,11 @@ func DecodePacket(packet Packet, expected_command_id PacketId, data any) error {
 	}
 }
 
-func EncodeHandshakePacket(value int) (Packet, error) {
+func EncodeHandshakePacket(value uint64) (Packet, error) {
 	return EncodePacket(PacketIdHandshake, &value)
 }
 
-func DecodeHandshakePacket(packet Packet) (result int, err error) {
+func DecodeHandshakePacket(packet Packet) (result uint64, err error) {
 	err = DecodePacket(packet, PacketIdHandshake, &result)
 	return
 }
